Failure handled

+++ main_.py
@@ -12,7 +12,12 @@ global counter
 
 counter = False
 
-def download_file(url,counter,location):
+with open('document1.csv', 'a') as fd:
+    writer = csv.writer(fd)
+    writer.writerow(["","Timestamp", "Server", "Tactic", "Latency", "Cost","Reliability",""])
+
+
+def download_file(url,location):
 
     try:
         starttime = time.time()
@@ -45,7 +50,10 @@ def download_file(url,counter,location):
 
         with open('document1.csv', 'a') as fd:
             writer = csv.writer(fd)
-            writer.writerow([dateTimeObj, flag, "1", (str)(final_time), psutil.cpu_percent(), reliability])
+            writer.writerow(["",dateTimeObj, flag, "1", (str)(final_time), psutil.cpu_percent(), reliability,""])
+
+        counter = False
+        return counter
 
 
     except Exception:
@@ -53,7 +61,7 @@ def download_file(url,counter,location):
         dateTimeObj = datetime.now()
         with open('document1.csv', 'a') as fd:
             writer = csv.writer(fd)
-            writer.writerow([dateTimeObj, location, "1", "-1", psutil.cpu_percent(), 0])
+            writer.writerow(["",dateTimeObj, location, "1", "-1", psutil.cpu_percent(), 0,""])
         return counter
 
 
@@ -101,7 +109,7 @@ def unzipfile(url):
 
         with open('document1.csv', 'a') as fd:
             writer = csv.writer(fd)
-            writer.writerow([dateTimeObj, flag, "2", (str)(endtime - starttime), psutil.cpu_percent(), reliability])
+            writer.writerow(["",dateTimeObj, flag, "2", (str)(endtime - starttime), psutil.cpu_percent(), reliability,""])
 
 
 def grep_file(url):
@@ -142,7 +150,7 @@ def grep_file(url):
         reliability = 1
         with open('document1.csv', 'a') as fd:
             writer = csv.writer(fd)
-            writer.writerow([dateTimeObj, flag, "3", (str)(endtime - starttime), psutil.cpu_percent(), reliability])
+            writer.writerow(["",dateTimeObj, flag, "3", (str)(endtime - starttime), psutil.cpu_percent(), reliability,""])
 
 
 def zip_file(url):
@@ -185,7 +193,7 @@ def zip_file(url):
         reliability = 1
         with open('document1.csv', 'a') as fd:
             writer = csv.writer(fd)
-            writer.writerow([dateTimeObj, flag, "4", (str)(endtime - starttime), psutil.cpu_percent(), reliability])
+            writer.writerow(["",dateTimeObj, flag, "3", (str)(endtime - starttime), psutil.cpu_percent(), reliability,""])
 
 
 def delete_file(url):
@@ -228,7 +236,7 @@ def delete_file(url):
         reliability = 1
         with open('document1.csv', 'a') as fd:
             writer = csv.writer(fd)
-            writer.writerow([dateTimeObj, flag, "5", (str)(endtime - starttime), psutil.cpu_percent(), reliability])
+            writer.writerow(["",dateTimeObj, flag, "3", (str)(endtime - starttime), psutil.cpu_percent(), reliability,""])
 
 
 pointer = True
@@ -239,62 +247,104 @@ while pointer == True:
     print("Location 1")
     location1 = 'http://ftp.utexas.edu/libreoffice/libreoffice/stable/6.2.8/deb/x86/LibreOffice_6.2.8_Linux_x86_deb.tar.gz'
     while flag == True:
-        # r = requests.get(location1)
-        # if r.status_code == 200:
-        # if r.status_code == 200:
-        counter = download_file(location1,counter,1)
-        if not counter:
-            unzipfile(location1)
-            grep_file(location1)
-            zip_file(location1)
-            delete_file(location1)
+
+        try:
+            r = requests.get(location1)
+            print ( r.status_code)
+            if r.status_code == 200:
+                counter = download_file(location1,1)
+                if not counter:
+                    unzipfile(location1)
+                    grep_file(location1)
+                    zip_file(location1)
+                    delete_file(location1)
+    
+            else:
+                dateTimeObj = datetime.now()
+                with open('document1.csv', 'a') as fd:
+                    writer = csv.writer(fd)
+                    writer.writerow(["", dateTimeObj, 1, "1", "-1", psutil.cpu_percent(), 0, ""])
+        except Exception:
+            print("Error")
         flag = False
         counter = False
 
-    print("Location 2")
-    flag = True
-    location2 = 'https://mirror.init7.net/tdf/libreoffice/stable/6.2.8/deb/x86/LibreOffice_6.2.8_Linux_x86_deb.tar.gz'
-    while flag == True:
-        # r = requests.get(location2)
-        # if r.status_code == 200:
-        # if r.status_code == 200:
-        counter = download_file(location2,counter,2)
-        if not counter:
-            unzipfile(location2)
-            grep_file(location2)
-            zip_file(location2)
-            delete_file(location2)
-        flag = False
-        counter = False
 
-    print("Location 3")
-    flag = True
-    location3 = 'http://tdf.mirror.rafal.ca/libreoffice/stable/6.2.8/deb/x86/LibreOffice_6.2.8_Linux_x86_deb.tar.gz'
-    while flag == True:
-        # r = requests.get(location3)
-        # if r.status_code == 200:
-        # if r.status_code == 200:
-        counter = download_file(location3,counter,3)
-        if not counter:
-            unzipfile(location3)
-            grep_file(location3)
-            zip_file(location3)
-            delete_file(location3)
-        flag = False
-        counter = False
+        print("Location 2")
+        flag = True
+        location2 = 'https://mirror.init7.net/tdf/libreoffice/stable/6.2.8/deb/x86/LibreOffice_6.2.8_Linux_x86_deb.tar.gz'
+        while flag == True:
+
+            try:
+                r = requests.get(location2)
+                print( r.status_code)
+                if r.status_code == 200:
+                    counter = download_file(location2, 1)
+                    if not counter:
+                        unzipfile(location2)
+                        grep_file(location2)
+                        zip_file(location2)
+                        delete_file(location2)
+                    
+                else:
+                    dateTimeObj = datetime.now()
+                    with open('document1.csv', 'a') as fd:
+                        writer = csv.writer(fd)
+                        writer.writerow(["", dateTimeObj, 2, "1", "-1", psutil.cpu_percent(), 0, ""])
+            except Exception:
+                print("Error")
+            flag = False
+            counter = False
+
+        print("Location 3")
+        flag = True
+        location3 = 'http://tdf.mirror.rafal.ca/libreoffice/stable/6.2.8/deb/x86/LibreOffice_6.2.8_Linux_x86_deb.tar.gz'
+        while flag == True:
+
+            try:
+                r = requests.get(location3)
+                print(r.status_code)
+                if r.status_code == 200:
+                    counter = download_file(location3, 1)
+                    if not counter:
+                        unzipfile(location3)
+                        grep_file(location3)
+                        zip_file(location3)
+                        delete_file(location3)
+               
+                else:
+                    dateTimeObj = datetime.now()
+                    with open('document1.csv', 'a') as fd:
+                        writer = csv.writer(fd)
+                        writer.writerow(["", dateTimeObj, 3, "1", "-1", psutil.cpu_percent(), 0, ""])
+            except Exception:
+                print("Error")
+            flag = False
+            counter = False
 
     print("Location 4")
     flag = True
     location4 = 'https://download.nus.edu.sg/mirror/tdf/libreoffice/stable/6.2.8/deb/x86/LibreOffice_6.2.8_Linux_x86_deb.tar.gz'
     while flag == True:
-        # r = requests.get(location4)
-        # if r.status_code == 200:
-        counter = download_file(location4,counter,4)
-        if not counter:
-            unzipfile(location4)
-            grep_file(location4)
-            zip_file(location4)
-            delete_file(location4)
+
+        try:
+            r = requests.get(location4)
+            print ( r.status_code)
+            if r.status_code == 200:
+                counter = download_file(location4,1)
+                if not counter:
+                    unzipfile(location4)
+                    grep_file(location4)
+                    zip_file(location4)
+                    delete_file(location4)
+             
+            else:
+                dateTimeObj = datetime.now()
+                with open('document1.csv', 'a') as fd:
+                    writer = csv.writer(fd)
+                    writer.writerow(["", dateTimeObj, 4, "1", "-1", psutil.cpu_percent(), 0, ""])
+        except Exception:
+            print("Error")
         flag = False
         counter = False
 
@@ -302,15 +352,25 @@ while pointer == True:
     flag = True
     location5 = 'https://tdf.c3sl.ufpr.br/libreoffice/stable/6.2.8/deb/x86/LibreOffice_6.2.8_Linux_x86_deb.tar.gz'
     while flag == True:
-        # r = requests.get(location5)
-        # if r.status_code == 200:
-        # if r.status_code == 200:
-        counter = download_file(location5,counter,5)
-        if not counter:
-            unzipfile(location5)
-            grep_file(location5)
-            zip_file(location5)
-            delete_file(location5)
+
+        try:
+            r = requests.get(location5)
+            print (r.status_code)
+            if r.status_code == 200:
+                counter = download_file(location5,1)
+                if not counter:
+                    unzipfile(location5)
+                    grep_file(location5)
+                    zip_file(location5)
+                    delete_file(location5)
+                
+            else:
+                dateTimeObj = datetime.now()
+                with open('document1.csv', 'a') as fd:
+                    writer = csv.writer(fd)
+                    writer.writerow(["", dateTimeObj, 5, "1", "-1", psutil.cpu_percent(), 0, ""])
+        except Exception:
+            print("Error")
         flag = False
         counter = False
 
@@ -318,15 +378,25 @@ while pointer == True:
     flag = True
     location6 = 'https://mirror-hk.koddos.net/tdf/libreoffice/stable/6.2.8/deb/x86/LibreOffice_6.2.8_Linux_x86_deb.tar.gz'
     while flag == True:
-        # r = requests.head(location5)
-        # if r.status_code == 200:
-        # if r.status_code == 200:
-        counter = download_file(location6,counter,6)
-        if not counter:
-            unzipfile(location6)
-            grep_file(location6)
-            zip_file(location6)
-            delete_file(location6)
+
+        try:
+            r = requests.get(location6)
+            print ( r.status_code)
+            if r.status_code == 200:
+                counter = download_file(location6,1)
+                if not counter:
+                    unzipfile(location6)
+                    grep_file(location6)
+                    zip_file(location6)
+                    delete_file(location6)
+                
+            else:
+                dateTimeObj = datetime.now()
+                with open('document1.csv', 'a') as fd:
+                    writer = csv.writer(fd)
+                    writer.writerow(["", dateTimeObj, 6, "1", "-1", psutil.cpu_percent(), 0, ""])
+        except Exception:
+            print("Error")
         flag = False
         counter = False
 
@@ -334,15 +404,25 @@ while pointer == True:
     flag = True
     location7 = 'http://mirrors.coreix.net/thedocumentfoundation/libreoffice/stable/6.2.8/deb/x86/LibreOffice_6.2.8_Linux_x86_deb.tar.gz'
     while flag == True:
-        # r = requests.get(location7)
-        # if r.status_code == 200:
-        # if r.status_code == 200:
-        counter = download_file(location7,counter,7)
-        if not counter:
-            unzipfile(location7)
-            grep_file(location7)
-            zip_file(location7)
-            delete_file(location7)
+
+        try:
+            r = requests.get(location7)
+            print (r.status_code)
+            if r.status_code == 200:
+                counter = download_file(location7,1)
+                if not counter:
+                    unzipfile(location7)
+                    grep_file(location7)
+                    zip_file(location7)
+                    delete_file(location7)
+                
+            else:
+                dateTimeObj = datetime.now()
+                with open('document1.csv', 'a') as fd:
+                    writer = csv.writer(fd)
+                    writer.writerow(["", dateTimeObj, 7, "1", "-1", psutil.cpu_percent(), 0, ""])
+        except Exception:
+            print("Error")
         flag = False
         counter = False
 
@@ -350,15 +430,26 @@ while pointer == True:
     flag = True
     location8 = 'https://mirror.aarnet.edu.au/pub/tdf/libreoffice/stable/6.2.8/deb/x86/LibreOffice_6.2.8_Linux_x86_deb.tar.gz'
     while flag == True:
-        # r = requests.get(location8)
-        # if r.status_code == 200:
-        # if r.status_code == 200:
-        counter = download_file(location8,counter,8)
-        if not counter:
-            unzipfile(location8)
-            grep_file(location8)
-            zip_file(location8)
-            delete_file(location8)
+
+        try:
+            r = requests.get(location8)
+            print (r.status_code)
+            if r.status_code == 200:
+                counter = download_file(location8,1)
+                if not counter:
+                    unzipfile(location8)
+                    grep_file(location8)
+                    zip_file(location8)
+                    delete_file(location8)
+                
+            else:
+                dateTimeObj = datetime.now()
+                with open('document1.csv', 'a') as fd:
+                    writer = csv.writer(fd)
+                    writer.writerow(["", dateTimeObj, 8, "1", "-1", psutil.cpu_percent(), 0, ""])
+            
+        except Exception:
+            print("Error")
         flag = False
         counter = False
